refactor: use logger for messages instead of console
The logs with console showed each time that the plugin was loaded into
Obsidian, to prevent this, I added a logger that checks if the build is
a dev one, or a production one. If it's a production one, the logs are
suppressed, and if it's a dev one, the logs are shown.